+++ plugins/voxelman/graphics/plugin.d
@@ -6,13 +6,14 @@ Authors: Andrey Penechko.
 
 module voxelman.graphics.plugin;
 
-import std.experimental.logger;
+import voxelman.log;
 import derelict.opengl3.gl3;
 import voxelman.container.buffer;
 import voxelman.math;
 import dlib.math.matrix;
 
 import pluginlib;
+import anchovy.iwindow;
 import anchovy.irenderer;
 import anchovy.shaderprogram;
 import voxelman.core.config;
@@ -29,35 +30,37 @@ shared static this()
 	pluginRegistry.regClientPlugin(new GraphicsPlugin);
 }
 
-string color_frag_shader = `
+string solid_frag_shader = `
 #version 330
-smooth in vec4 theColor;
-out vec4 outputColor;
-const vec4 fogcolor = vec4(0.6, 0.8, 1.0, 1.0);
-const float fogdensity = .00001;
-vec4 gamma(vec4 color) {
-	return vec4(pow(color.xyz, vec3(1.0/2.0)), color.a);
-}
+smooth in vec4 frag_color;
+out vec4 out_color;
+
 void main() {
-	float z = gl_FragCoord.z / gl_FragCoord.w;
-	float fogModifier = clamp(exp(-fogdensity * z * z), 0.0, 1);
-	outputColor = (mix(fogcolor, theColor, fogModifier));
+	out_color = frag_color;
 }
 `;
 
 string color_frag_shader_transparent = `
 #version 330
-smooth in vec4 theColor;
-out vec4 outputColor;
-const vec4 fogcolor = vec4(0.6, 0.8, 1.0, 1.0);
-const float fogdensity = .00001;
-vec4 gamma(vec4 color) {
-	return vec4(pow(color.xyz, vec3(1.0/2.0)), color.a);
-}
+smooth in vec4 frag_color;
+out vec4 out_color;
+
 void main() {
-	float z = gl_FragCoord.z / gl_FragCoord.w;
-	float fogModifier = clamp(exp(-fogdensity * z * z), 0.0, 1);
-	outputColor = (vec4(mix(fogcolor, theColor, fogModifier).xyz, 0.5));
+	out_color = vec4(frag_color.xyz, 0.5);
+}
+`;
+
+string vert_shader_2d = `
+#version 330
+layout(location = 0) in vec2 position;
+layout(location = 1) in vec4 color;
+
+smooth out vec4 frag_color;
+uniform mat4 proj_mat;
+
+void main() {
+	gl_Position = proj_mat * vec4(position.xy, 0, 1);
+	frag_color = color;
 }
 `;
 
@@ -68,10 +71,10 @@ layout(location = 1) in vec4 color;
 uniform mat4 projection;
 uniform mat4 view;
 uniform mat4 model;
-smooth out vec4 theColor;
+smooth out vec4 frag_color;
 void main() {
 	gl_Position = projection * view * model * position;
-	theColor = color;
+	frag_color = color;
 }
 `;
 
@@ -82,12 +85,10 @@ layout(location = 1) in vec4 color;
 uniform mat4 projection;
 uniform mat4 view;
 uniform mat4 model;
-smooth out vec4 theColor;
+smooth out vec4 frag_color;
 void main() {
 	gl_Position = projection * view * model * (position);
-	//vec4 unpackedColor = vec4(1,1,1,1);
-	//vec4 unpackedColor = vec4((color & 31)/31, ((color >> 5) & 31)/31, ((color >> 10) & 31)/31, 1);
-	theColor = color;
+	frag_color = color;
 }
 `;
 
@@ -97,22 +98,28 @@ private:
 	uint vao;
 	uint vbo;
 	EventDispatcherPlugin evDispatcher;
+	float[4][4] ortho_projection;
 
 public:
 	FpsCamera camera;
 	Batch debugBatch;
+	Batch2d overlayBatch;
 
 	ShaderProgram chunkShader;
 	ShaderProgram solidShader;
+	ShaderProgram solidShader2d;
 	ShaderProgram transChunkShader;
 
 	IRenderer renderer;
+	IWindow window;
+
 	ConfigOption cameraSensivity;
 	ConfigOption cameraFov;
 
-	GLuint projectionLoc = 2; //perspective
-	GLuint viewLoc = 3; //camera trandformation
-	GLuint modelLoc = 4; //model transformation
+	GLint projectionLoc = 2; //perspective
+	GLint viewLoc = 3; //camera trandformation
+	GLint modelLoc = 4; //model transformation
+	GLint projLoc = 5; // 2d projection
 
 
 	mixin IdAndSemverFrom!(voxelman.graphics.plugininfo);
@@ -140,38 +147,34 @@ public:
 		evDispatcher.subscribeToEvent(&draw);
 
 		renderer = gui.renderer;
+		window = gui.window;
 
 		glGenVertexArrays(1, &vao);
 		glGenBuffers( 1, &vbo);
 
 		// Setup shaders
-		chunkShader = renderer.createShaderProgram(chunk_vert_shader, color_frag_shader);
+		chunkShader = renderer.createShaderProgram(chunk_vert_shader, solid_frag_shader);
 		transChunkShader = renderer.createShaderProgram(chunk_vert_shader, color_frag_shader_transparent);
-		solidShader = renderer.createShaderProgram(solid_vert_shader, color_frag_shader);
+		solidShader = renderer.createShaderProgram(solid_vert_shader, solid_frag_shader);
+		solidShader2d = renderer.createShaderProgram(vert_shader_2d, solid_frag_shader);
 
-		chunkShader.bind;
-			modelLoc = glGetUniformLocation( solidShader.handle, "model" );//model transformation
-			viewLoc = glGetUniformLocation( solidShader.handle, "view" );//camera trandformation
-			projectionLoc = glGetUniformLocation( solidShader.handle, "projection" );//perspective
-
-			glUniformMatrix4fv(modelLoc, 1, GL_FALSE,
-				cast(const float*)Matrix4f.identity.arrayof);
-			glUniformMatrix4fv(viewLoc, 1, GL_FALSE,
-				cast(const float*)camera.cameraMatrix);
-			glUniformMatrix4fv(projectionLoc, 1, GL_FALSE,
-				cast(const float*)camera.perspective.arrayof);
-		chunkShader.unbind;
+		modelLoc = glGetUniformLocation( solidShader.handle, "model" ); // model transformation
+		viewLoc = glGetUniformLocation( solidShader.handle, "view" ); // camera trandformation
+		projectionLoc = glGetUniformLocation( solidShader.handle, "projection" ); // perspective
+		projLoc = glGetUniformLocation( solidShader2d.handle, "proj_mat" ); // 2d projection
 	}
 
 	override void postInit()
 	{
 		renderer.setClearColor(165,211,238);
 		camera.aspect = cast(float)renderer.framebufferSize.x/renderer.framebufferSize.y;
+		updateOrtoMatrix();
 	}
 
 	private void onWindowResizedEvent(ref WindowResizedEvent event)
 	{
 		camera.aspect = cast(float)event.newSize.x/event.newSize.y;
+		updateOrtoMatrix();
 	}
 
 	void resetCamera()
@@ -184,8 +187,6 @@ public:
 
 	private void draw(ref RenderEvent event)
 	{
-		glScissor(0, 0, renderer.framebufferSize.x, renderer.framebufferSize.y);
-		glViewport(0, 0, renderer.framebufferSize.x, renderer.framebufferSize.y);
 		glClear(GL_COLOR_BUFFER_BIT|GL_DEPTH_BUFFER_BIT);
 		glEnable(GL_DEPTH_TEST);
 
@@ -201,8 +202,14 @@ public:
 		evDispatcher.postEvent(RenderTransparent3dEvent(renderer));
 
 		renderer.enableAlphaBlending();
+
 		evDispatcher.postEvent(Render2Event(renderer));
+
+		draw(overlayBatch);
+		overlayBatch.reset();
+
 		evDispatcher.postEvent(Render3Event(renderer));
+
 		renderer.disableAlphaBlending();
 		renderer.flush();
 	}
@@ -210,34 +217,55 @@ public:
 	void draw(Batch batch)
 	{
 		solidShader.bind;
-		drawBuffer(batch.triBuffer.data, GL_TRIANGLES);
-		drawBuffer(batch.lineBuffer.data, GL_LINES);
-		drawBuffer(batch.pointBuffer.data, GL_POINTS);
-		solidShader.unbind;
-	}
-
-private:
-
-	void drawBuffer(ColoredVertex[] buffer, uint mode)
-	{
-		if (buffer.length == 0) return;
 
 		glUniformMatrix4fv(modelLoc, 1, GL_FALSE, cast(const float*)Matrix4f.identity.arrayof);
 		glUniformMatrix4fv(viewLoc, 1, GL_FALSE, camera.cameraMatrix);
 		glUniformMatrix4fv(projectionLoc, 1, GL_FALSE, cast(const float*)camera.perspective.arrayof);
+
+		drawBuffer(batch.triBuffer.data, GL_TRIANGLES);
+		drawBuffer(batch.lineBuffer.data, GL_LINES);
+		drawBuffer(batch.pointBuffer.data, GL_POINTS);
+
+		solidShader.unbind;
+	}
+
+	void draw(Batch2d batch)
+	{
+		solidShader2d.bind;
+
+		glUniformMatrix4fv(projLoc, 1, GL_FALSE, &ortho_projection[0][0]);
+
+		drawBuffer(batch.triBuffer.data, GL_TRIANGLES);
+		drawBuffer(batch.lineBuffer.data, GL_LINES);
+		drawBuffer(batch.pointBuffer.data, GL_POINTS);
+
+		solidShader2d.unbind;
+	}
+
+private:
+
+	void updateOrtoMatrix()
+	{
+		auto w = renderer.framebufferSize.x;
+		auto h = renderer.framebufferSize.y;
+		ortho_projection =
+		[
+			[ 2f/w, 0f,   0f, 0f ],
+			[ 0f,  -2f/h, 0f, 0f ],
+			[ 0f,   0f,  -1f, 0f ],
+			[-1f,   1f,   0f, 1f ],
+		];
+	}
+
+	void drawBuffer(VertexType)(VertexType[] buffer, uint mode)
+	{
+		if (buffer.length == 0) return;
 		glBindVertexArray(vao);
 		glBindBuffer(GL_ARRAY_BUFFER, vbo);
-		glBufferData(GL_ARRAY_BUFFER, buffer.length*ColoredVertex.sizeof, buffer.ptr, GL_DYNAMIC_DRAW);
-		glEnableVertexAttribArray(0);
-		glEnableVertexAttribArray(1);
-		// positions
-		glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, ColoredVertex.sizeof, null);
-		// color
-		glVertexAttribPointer(1, 3, GL_UNSIGNED_BYTE, GL_TRUE, ColoredVertex.sizeof, cast(void*)(12));
+		glBufferData(GL_ARRAY_BUFFER, buffer.length*VertexType.sizeof, buffer.ptr, GL_DYNAMIC_DRAW);
+		VertexType.setAttributes();
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
-
 		glDrawArrays(mode, 0, cast(uint)(buffer.length));
-
 		glBindVertexArray(0);
 	}
 }