+++ source/voxelman/graphics/color.d
@@ -7,23 +7,23 @@ module voxelman.graphics.color;
 
 import voxelman.math : Vector;
 
-alias Color3ub = Vector!(ubyte, 3);
+alias Color4ub = Vector!(ubyte, 4);
 
-enum Colors : Color3ub
+enum Colors : Color4ub
 {
-	black = Color3ub(0, 0, 0),
-	white = Color3ub(255, 255, 255),
-	red = Color3ub(255, 0, 0),
-	green = Color3ub(0, 255, 0),
-	blue = Color3ub(0, 0, 255),
-	cyan = Color3ub(0, 255, 255),
-	magenta = Color3ub(255, 0, 255),
-	yellow = Color3ub(255, 255, 0),
-	gray = Color3ub(128, 128, 128),
+	black = Color4ub(0, 0, 0, 255),
+	white = Color4ub(255, 255, 255, 255),
+	red = Color4ub(255, 0, 0, 255),
+	green = Color4ub(0, 255, 0, 255),
+	blue = Color4ub(0, 0, 255, 255),
+	cyan = Color4ub(0, 255, 255, 255),
+	magenta = Color4ub(255, 0, 255, 255),
+	yellow = Color4ub(255, 255, 0, 255),
+	gray = Color4ub(128, 128, 128, 255),
 }
 
 
-Color3ub[] colorsArray = [
+Color4ub[] colorsArray = [
 	Colors.black, Colors.white, Colors.red,
 	Colors.green, Colors.blue, Colors.cyan,
 	Colors.magenta, Colors.yellow